Decouple shared metadata element rendering
Move metadata and viewport element construction into a neutral metadata-elements module. Both the legacy and parallel component implementations now import the shared renderer, so the parallel path no longer depends on the legacy metadata component and either implementation can be removed independently.\n\nPreserve explicit coverage for the disabled-flag behavior where notFound() from a slot's generateMetadata reaches the custom not-found boundary. Update the server trace snapshot for the extracted module.\n\n<!-- NEXT_JS_LLM -->
